recreate materialized views concurrently
We were making other queries wait for the refresh query to finish. This query was taking longer and longer, and this led to very long queues of requests to player pages.

Running REFRESH MATERIALIZED VIEW with CONCURRENTLY means that we create a new version of the view as a separate entity, and lock the original view only for a brief time when we replace it with the new version.

https://www.postgresql.org/docs/15/sql-refreshmaterializedview.html

To make this work, we have to split refresh_view and create_view: you can only run refresh concurrently after you have some data, so the first run has to be a regular refresh materialized view.