cache models during startup
Models basically never change, so there is no need to go the database more than once.

Models can be changed by reindexing, so we explicitly reload this cache in ModelIndexer now.

Team release page might not sound like an ideal place for this test, but it’s our home page, so this means something like “when we load a page, we don’t query the models table, and it’s enough to test this for a default page”.